+++ Book_Exchange_System/Program.cs
@@ -16,7 +16,7 @@ namespace Book_Exchange_System
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Applicants());
+            Application.Run(new Admin());
         }
     }
 }
